Split private mirror functions for reuse
These files will be copied for now and then refactored later into
reusable packages.
